extract MeshReplicatorConnectionParams into a separate file
